Post-merge fixes
* Fixed duplicate source files in iOS target in pbxproj file
* Removed obsolete source files that reappeared in the merge
* Fixed 32-bit compile warning in CBLView.mm
* Fixed submodule URLs

+++ Source/API/CBLView.mm
@@ -469,7 +469,7 @@ static id<CBLViewCompiler> sCompiler;
 
 
 - (NSUInteger) totalRows {
-    return self.index->rowCount();
+    return (NSUInteger) self.index->rowCount();
 }
